lcow: Allow the client to add or remove capabilities
